requirements.txt
ccxt
requests
win10toast
pushbullet.py
twilio